reorganize commit code
